requirements.txt
fastapi
uvicorn
pydantic
google-analytics-data
google-auth
openai
